Refactor: Move GraphQLTypeAnnotation out of the definition of Schema and into the type parameters.
This is part of a larger change to move away from GraphQLTypeAnnotation in Schema. The overall goal is to model types more sanely in the Schema. In particular, the fact that the "default" is a nullable type makes things a lot harder, because things are subtractive instead of additive.

This also locks us into a very GraphQL-centric way of thinking. But we should move away from GraphQL-focused thinking as much as possible.