When setting a source in a Database, avoid incrementing the epoch if the source is identical to the previously set source.
We can attempt to re-set an identical source if the user hits cmd+s on an unchanged file. Our notifier will receive some events, with no indication that the file did not change. The natural place to determine whether the file is unchanged is the Database, which knows about the previous value.

If we need a version of this function that avoids the equality check (e.g. if you are sure that the source actually has changed, e.g. you mutated the previous one, it's an incrementing counter, etc.) then we can add another message