Rename ServerSelectableFieldId to ServerEntityId
In terms of naming, I think the fundamental entities in Isograph are:

- selections
- selectable (formerly known as field). In user-facing-land these are fields and pointers
- entity (scalars and objects)

They can all be prefixed with server or client, and scalar or object

The goal is that selections contain ids of selectables which contain ids of entities, and that this is a scalar or object is preserved throughout all of that.  i.e. a ScalarSelection contains a ScalarSelectableId, which contains a ScalarEntityId (for the target)

I am questioning whether Server vs Client is an appropriate distinction to have in the types. There's never a situation in which we care, except for when creating merged selection sets (which must contain only server selectables). Regardless, it will probably stay for now.