Refactor: ServerScalarSelectable does not have a TClientFieldVariableDefinitionAssociatedData type param
Instead, it always has arguments with variables with inner type ServerEntityId

We can do this because we are always processing fields (ServerScalarSelectable) after all server scalars and objects have been processed. So, when we create the ServerScalarSelectable, we can look up the argument's actual type, and emit an error if that type is not defined.

We no longer need to do this in a later validation step!

This is part of a larger refactor #471