Memoize create_schema (the first part)
Note that this means that we no longer re-parse the GraphQL schema and recreate that unvalidated schema if only a JS file has changed!!!!

On the pet demo, locally, this worsens perf, we actually clone the schema, which I guess is slower than just redoing that work :sob: But that's okay. We'll break the schema up into smaller parts and avoid needing such a massive clone in the future.

At this point, we are free and clear to start work on the language server!!!!!!