resolve_position: use a generic type parameter in resolve
- Basically, we used to do

```
let parent_path = ...;
// move parent_path into child_parent
let child_parent = ChildParent::Child(parent_path);
do {
  // move child_parent into child.resolve, extract it back out with ?
  child_parent = child.resolve(child_parent, ...)?;
}
// extract back parent_path
let parent_path = child_parent.assert_child().parent;
```

This is bad! The child.resolve could have returned a ChildParent::Parent, in which case we would panic. Yikes!

So, instead, we allow child.resolve to accept the parent_path directly. But! child.resolve is also called with child_path. So, instead, we must accept a generic (`P: Into<ChildParent>`) and accept that.

This PR does that. Now things are glorious.

------

Is this the trait's final form? Maybe not. It would be nice to extract the redundant span.contains logic, a la:

```
// param is WithSpan<T: ResolvePosition>

if param.span.contains(position) {
  return ControlFlow::Continue(into_parent)
} else {
  return ControlFlow::Break(
    param.item.resolve_unconditional(into_parent.into())
  )
}
```

But that is a task for another day.