use isograph database fields for source files (#646)
`IsographDatabase` now has tracked fields on it: 
```
    #[tracked]
    iso_literal_map: IsoLiteralMap,
    #[tracked]
    standard_sources: StandardSources,
    #[tracked]
    open_file_map: OpenFileMap,
```

Access to these fields is gated via methods like `db.get_open_file_map` and `get_open_file_map_mut`, which expose tracked and untracked accessors. (The `_mut` methods expose only a tracked accessor.)

Tracked:
```
/// Provides tracked access to the underlying field. On a tracked map, tracked access is
/// required for correctness if adding or removing an item from the underlying map
/// affects the result. For example, when iterating, use tracked access.
```

Untracked:

```
/// Provides untracked access to the underlying field. On a tracked map, untracked access is
/// only correct if every accessed value is later tracked, and unrelated additions to or
/// removals from the underlying map do not affect the result. For example, if the map's values
/// are themselves [`SourceId`](SourceId) and you are accessing a single, specific value by
/// key, it is safe to use an untracked access.
///
/// This may improve performance, as it means that unrelated additions and removals to the map
/// will not invalidate this function call in pico.
```

### Probably a performance boost

This should provide a performance boost. For example, if we:

- go to definition on a selection in file 1 (which goes through an untracked access of open_file_map)
- modify open file 2 (modifying open_file_map)
- go to definition on that same selection

If the access in step 1 was tracked, we would have to revalidate more work during step 1, i.e. we would have to reread the file, meaning that everything upstream (e.g. parsing the iso literal etc) would have to revalidate. But instead, it's untracked (i.e. meaning that we're asserting that changes to unrelated keys don't matter), we find that the value is unchanged, and we can revalidate simply by seeing that file 1 has not been modified since we last checked it.