call subscription callback if data changed before commit (#731)
subscribe is called in an effect. (We should actually subscribe during the
initial render.) Because it's called in an effect, we might have missed some
changes since the initial render! So, when we initially subscribe, we re-read and call the
subscription (i.e. re-render) if the fragment data has changed.