Rename artifact look cache to file system state
Rename it to FileSystemState, modifies it to map the filepaths to a
tuple the file hash and the content as string. Also, remove it's public
method that writes to disk and add a methdo to handle to compare its
keys with other state keys in order to write/delete to/from file system
the proper artifacts.