Write only changed files to disk in watch mode
First of all, keep the file system state in the CompilerState and pass
the state to the compile function. A new struct ChangedArtifacts keeps
all the artifacts that will be added and deleted. the function
get_artifacts_to_write returns the changed artifacts based on the new
and old file system states.