Massive commit: embedded locations + lack of locations
This commit takes a giant, disorganized step towards two goals

- Entities and selectables do not location info on them. Only declarations have location info.
    - Instead, there is a selectable_definition_location fn for goto_definition etc
    - We didn't fully achieve this goal! There is still location info in e.g. selection sets. But we should have two separate data structures: selection sets in declarations and selection sets in selectables. It's probably the case that we want to have structs that are generic over TLocation, which is EmbeddedLocation in declarations and () otherwise.
    - We need to do more, e.g. have the ability to look up the location of an argument definition, so that you can go to def on an argument definition on a field.

- Everything we parse has an embedded location
    - Probably achieved?

We're closer to getting rid of WithSpan.

I will do another pass - we arbitrarily chose EmbeddedLocation's as the target for everything. And that's awkward, we have a few places where we create fake EmbeddedLocations. These probably panic if we try to print them. Maybe that second pass will add the TLocation param?